IMPALA-4875: Bound maximum statestore topic update size to 2GB.
This patch enforces a maximum topic update size for the statestore to be
2GB. If that limit is exceeded, the statestore will split topic updates
over several RPCs to the subscriber. If a single topic key + value
exceeds the limit, it will be dropped by the statestore when the
subscriber tries to publish it.

Doing so introduces the possibility that a catalog update (all updates
pertaining to a single catalog version) might be split across two
separate UpdateState() RPCs. To ensure that all updates are flushed to
the catalog FE at the same time, we buffer updates until a new object of
type CATALOG is received.

Remove the special treatment of deletions in the UpdateState()
RPC. Previously they were not sent with their version.

New statestore tests added to ensure:

* Update sizes don't exceed the configured max.
* 'from' version is properly handled (not previously tested) so that
  subscribers can process only a prefix of the update log.

TODO : Check min subscriber version

Change-Id: Icbe42704d3c98d9619e3737786b75ee67fb2a264